accept torrent file from stdin
This accepts a torrent-file from stdin downloading or seeding. Before,
this was already possible with the info and ls command, this extends
this support to cover all operations.

The dependency on `parse-torrent` has been bumped to a version that
actually throws errors. It was previously wrapped in a try catch even
though it returned `null` instead of throwing an error.